pyproject.toml
[project]
name = "oscar-ballot"
version = "0.1.0"
description = "Oscar ballot predictions using LLM voters and judge."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "pydantic>=2.7.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.ruff]
line-length = 130
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
